revamp blog section and polish
