+++ app/home.tsx
@@ -4,7 +4,7 @@ import dynamic from "next/dynamic";
 import Intro from "@/components/Intro";
 
 const About = dynamic(() => import("@/components/About"));
-const Exp = dynamic(() => import("@/components/Exp"));
+const Blog = dynamic(() => import("@/components/Blog"));
 const Projects = dynamic(() => import("@/components/Projects"));
 const LatestCode = dynamic(() => import("@/components/LatestCode"));
 
@@ -13,7 +13,7 @@ export default function Home() {
     <>
       <Intro />
       <About />
-      <Exp />
+      <Blog />
       <Projects />
       <LatestCode />
     </>
