+++ constants/constants.ts
@@ -26,44 +26,37 @@ export const expInfo = [
     logo: "/palantir/palantir.jpg",
     company: "Palantir",
     description: "Forward deployed in London",
-    bg: "#a0a0a0",
-    new: false,
+    type: "Internship",
   },
   {
     logo: "/ethglobal/ethglobal.png",
     company: "ETHGlobal",
     description: "Travelling the world with ETH",
-    bg: "#be889a",
-    new: true,
+    type: "Internship",
   },
   {
     logo: "/faire/faire.png",
     company: "Faire",
     description: "A day in the life at Faire",
-    bg: "#696969",
-    new: true,
+    type: "Internship",
   },
-
   {
     logo: "/hackthenorth/hackthenorth.png",
     company: "Hack the North",
     description: "Frontend at Canada's biggest hackathon",
-    bg: "#7a7ea1",
-    new: false,
+    type: "Volunteering",
   },
   {
     logo: "/uwblueprint/bp.png",
     company: "UW Blueprint",
     description: "Helping out the Children's Aid Society",
-    bg: "#5a83bf",
-    new: false,
+    type: "Volunteering",
   },
   {
     logo: "/civilia/civilia.jpeg",
     company: "Civilia",
     description: "First co-op jitters",
-    bg: "#8ba16a",
-    new: false,
+    type: "Internship",
   },
 ];
